requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
python-dateutil==2.8.2